WIP 6: Oncall Schedule
Done:
Generate Plan does the best it can. If multiple possibilities exist
for some person post generation, it depends on searching to resolve
them. There are tests for `generate-plan` and the overall code is
cleaner.

Pending:
1. Allow a way to assist someone in their oncall. This can be done by
merging constraints. For example, if I want Neha to be present when
Pranav is oncall, I only need to merge Neha's constraints into
Pranav's.
2. Add tests for `solve-by-search`
3. Clean up the code. Fix documentation and variables.