Rewrite of the tic-tac-toe game
The code is in a much better shape now, though it still feels
completely brute-force to me. Not much elegance in the detection
approaches.

I've removed the refs altogether. There might be good value in refs
when computers come along, but we'll get to it when we get to it.

I've added property tests for a bunch of things

I've cleaned up the code.